requirements.txt
requests>=2.31.0
yfinance>=0.2.36
pandas>=2.1.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=7.4.0
pytest-mock>=3.12.0
pydantic>=2.5.0
pyyaml>=6.0.1
numpy>=1.26.0
streamlit>=1.32.0
